test(cleanup): align scanner coverage tests
