+++ tests/orchestration/test_overlay_selection_label_domain.py
@@ -20,11 +20,13 @@ from frame_compare.vs.types import HDRMetadata
 
 
 class FakeFFmpegRunner:
-    def extract_frame(self, _video: Path, _frame_num: int, output: Path) -> None:
+    def extract_frame(self, video: Path, frame_num: int, output: Path) -> None:
+        _, _ = video, frame_num
         output.parent.mkdir(parents=True, exist_ok=True)
         Image.new("RGB", (10, 10), color=(0, 0, 0)).save(output, format="PNG")
 
-    def probe_hdr(self, _video: Path) -> HDRMetadata | None:
+    def probe_hdr(self, video: Path) -> HDRMetadata | None:
+        _ = video
         return HDRMetadata(
             mastering_display=None,
             max_cll=None,
