+++ tests/services/test_metadata.py
@@ -1,12 +1,9 @@
 from collections.abc import AsyncIterator
-from typing import get_args
 
 import httpx
 import pytest
 import respx
 
-import frame_compare.services.metadata_parsing as metadata_parsing
-import frame_compare.services.tmdb_lookup as tmdb_lookup
 from frame_compare.services.errors import MetadataError, TmdbError, TmdbRateLimitedError
 from frame_compare.services.metadata import lookup_tmdb, parse_filename, resolve_metadata
 from frame_compare.services.types import MetadataConfig, ParsedMetadata, TmdbMetadata
@@ -157,187 +154,7 @@ def test_parse_filename_parsers_raise_falls_back_to_stem(mocker) -> None:
     )
 
 
-def test_metadata_parsing_direct_module_merges_guessit_and_anitopy_fallback(mocker) -> None:
-    """Direct parser module keeps primary fields while filling missing alternate fields."""
-    mocker.patch(
-        "frame_compare.services.metadata_parsing.guessit",
-        return_value={
-            "title": "Movie.Name",
-            "year": "2024",
-            "screen_size": "2160p",
-        },
-    )
-    mocker.patch(
-        "frame_compare.services.metadata_parsing.anitopy.parse",
-        return_value={
-            "anime_title": "Alternate Title",
-            "anime_episode": "07",
-            "release_group": "SubsPlease",
-        },
-    )
-
-    result = metadata_parsing.parse_filename("Movie.Name.2024.E07.2160p.mkv")
-
-    assert result == ParsedMetadata(
-        title="Movie Name",
-        year=2024,
-        episode=7,
-        release_group="SubsPlease",
-        resolution="2160p",
-    )
-    assert parse_filename is metadata_parsing.parse_filename
-
-
-def test_metadata_parsing_direct_module_uses_anitopy_first_for_bracketed_names(
-    mocker,
-) -> None:
-    calls: list[str] = []
-
-    def fake_anitopy_parse(filename: str) -> dict[str, object]:
-        calls.append(f"anitopy:{filename}")
-        return {
-            "anime_title": "Anime_Title",
-            "anime_episode": 3,
-            "release_group": "Group",
-        }
-
-    def fake_guessit(filename: str) -> dict[str, object]:
-        calls.append(f"guessit:{filename}")
-        return {"title": "Western Title", "year": 2022}
-
-    mocker.patch("frame_compare.services.metadata_parsing.anitopy.parse", fake_anitopy_parse)
-    mocker.patch("frame_compare.services.metadata_parsing.guessit", fake_guessit)
-
-    result = metadata_parsing.parse_filename("[Group] Anime_Title - 03.mkv")
-
-    assert calls == [
-        "anitopy:[Group] Anime_Title - 03.mkv",
-        "guessit:[Group] Anime_Title - 03.mkv",
-    ]
-    assert result.title == "Anime Title"
-    assert result.year == 2022
-    assert result.episode == 3
-    assert result.release_group == "Group"
-
-
 # ─── TMDB Lookup Tests ────────────────────────────────────────────────────────
-
-
-def test_tmdb_metadata_media_type_is_closed_domain() -> None:
-    assert set(get_args(TmdbMetadata.__dataclass_fields__["media_type"].type)) == {"movie", "tv"}
-
-
-@pytest.mark.anyio
-async def test_tmdb_lookup_direct_module_maps_request_and_response_with_mock_transport() -> None:
-    requests: list[httpx.Request] = []
-
-    def handler(request: httpx.Request) -> httpx.Response:
-        requests.append(request)
-        assert request.url.host == "api.themoviedb.org"
-        assert request.url.params["api_key"] == "b" * 32
-        assert request.url.params["query"] == "Arrival"
-        assert request.url.params["year"] == "2016"
-        assert request.url.params["include_adult"] == "false"
-        return httpx.Response(
-            200,
-            json={
-                "results": [
-                    {
-                        "id": 329865,
-                        "title": "Arrival",
-                        "original_title": "Arrival",
-                        "release_date": "2016-11-10",
-                        "media_type": "movie",
-                        "poster_path": "/poster.jpg",
-                        "backdrop_path": None,
-                    },
-                    {
-                        "id": 123,
-                        "name": "Ignored Person",
-                        "media_type": "person",
-                    },
-                ]
-            },
-        )
-
-    async with httpx.AsyncClient(transport=httpx.MockTransport(handler)) as client:
-        result = await tmdb_lookup.lookup_tmdb(
-            ParsedMetadata(title="Arrival", year=2016),
-            MetadataConfig(api_key="b" * 32, timeout_seconds=3.5),
-            client,
-        )
-
-    assert requests
-    assert result == TmdbMetadata(
-        tmdb_id=329865,
-        title="Arrival",
-        original_title="Arrival",
-        year=2016,
-        media_type="movie",
-        poster_url="https://image.tmdb.org/t/p/original/poster.jpg",
-        backdrop_url=None,
-    )
-    assert lookup_tmdb is tmdb_lookup.lookup_tmdb
-
-
-@pytest.mark.anyio
-async def test_tmdb_lookup_direct_module_classifies_http_and_transport_errors() -> None:
-    async with httpx.AsyncClient(
-        transport=httpx.MockTransport(lambda request: httpx.Response(401))
-    ) as client:
-        with pytest.raises(TmdbError) as excinfo:
-            await tmdb_lookup.lookup_tmdb(
-                ParsedMetadata(title="Arrival"),
-                MetadataConfig(api_key="c" * 32),
-                client,
-            )
-
-    assert excinfo.value.context.details == {"reason": "Invalid API key"}
-
-    def raise_connect_error(request: httpx.Request) -> httpx.Response:
-        raise httpx.ConnectError(f"leaked-key={'d' * 32}", request=request)
-
-    async with httpx.AsyncClient(transport=httpx.MockTransport(raise_connect_error)) as client:
-        with pytest.raises(TmdbError) as excinfo:
-            await tmdb_lookup.lookup_tmdb(
-                ParsedMetadata(title="Arrival"),
-                MetadataConfig(api_key="d" * 32),
-                client,
-            )
-
-    assert excinfo.value.context.details == {"reason": "Request failed"}
-    assert "d" * 32 not in excinfo.value.context.message
-
-
-@pytest.mark.anyio
-async def test_tmdb_lookup_direct_module_classifies_rate_limit_and_timeout() -> None:
-    async with httpx.AsyncClient(
-        transport=httpx.MockTransport(lambda request: httpx.Response(429))
-    ) as client:
-        with pytest.raises(TmdbRateLimitedError):
-            await tmdb_lookup.lookup_tmdb(
-                ParsedMetadata(title="Arrival"),
-                MetadataConfig(api_key="e" * 32),
-                client,
-            )
-
-    def raise_timeout(request: httpx.Request) -> httpx.Response:
-        raise httpx.TimeoutException("too slow", request=request)
-
-    async with httpx.AsyncClient(transport=httpx.MockTransport(raise_timeout)) as client:
-        with pytest.raises(TmdbError) as excinfo:
-            await tmdb_lookup.lookup_tmdb(
-                ParsedMetadata(title="Arrival"),
-                MetadataConfig(api_key="f" * 32),
-                client,
-            )
-
-    assert excinfo.value.context.details == {"reason": "Request timed out"}
-
-
-def test_tmdb_lookup_direct_module_validates_api_key_shape() -> None:
-    assert tmdb_lookup.is_valid_tmdb_api_key("0123456789abcdefABCDEF0123456789")
-    assert not tmdb_lookup.is_valid_tmdb_api_key("g" * 32)
 
 
 @pytest.mark.anyio
